% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChisqInd.R
\name{ChisqInd}
\alias{ChisqInd}
\title{Chi Square Test of Independence}
\usage{
ChisqInd(observed_matrix)
}
\arguments{
\item{observed_matrix}{A rectangular matrix of observed values}
}
\value{
a list containing 3 values
\itemize{
    \item \code{expected_matrix} A matrix with equivalent dimensions
    of the observed containing the expected counts
    \item \code{ts} The chi square test statistic
    \item \code{p} The chi square test p-value
}
}
\description{
Given an observed rectangular matrix, retreive the test statistic
and p-value for the chi-squared test of independence
}
\examples{
# testing 2X2 matrix
matrix1 <- matrix(c(10, 20, 15, 25), nrow = 2, byrow = TRUE)
result <- ChisqInd(matrix1)
print( result$expected_matrix )
print( result$ts )
print( result$p )
}
